Add --no-verify flag for skipping TLS verification.
Passing --no-verify will skip TLS verification on any HTTP requests
made. Sites with broken TLS can still contain interesting info.

This adds urllib3 to the requirements.txt file because we need
to use it to disable insecure request warnings when --no-verify
is used. This does not actually add a new dependency however since
the requests package already depends on it.